adds test and query fix

+++ src/test/java/io/supertokens/test/userRoles/UserRolesStorageTest.java
@@ -17,6 +17,7 @@
 package io.supertokens.test.userRoles;
 
 import io.supertokens.ProcessState;
+import io.supertokens.emailverification.User;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
@@ -578,6 +579,36 @@ public class UserRolesStorageTest {
             String[] userRoles = storage.getRolesForUser(userId);
             assertEquals(0, userRoles.length);
         }
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testRetrievingUsersForRoles() throws Exception {
+        String[] args = { "../" };
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
+            return;
+        }
+        UserRolesSQLStorage storage = StorageLayer.getUserRolesStorage(process.main);
+
+        // create a role
+        String role = "role";
+        UserRoles.createNewRoleOrModifyItsPermissions(process.main, role, null);
+
+        // add role to multiple users
+        String[] userIds = new String[] { "user1", "user2", "user3" };
+        for (String userId : userIds) {
+            UserRoles.addRoleToUser(process.main, userId, role);
+        }
+
+        String[] userIdsWithSameRole = storage.getUsersForRole(role);
+
+        // check that the users you have the role is correct
+        Utils.checkThatArraysAreEqual(userIds, userIdsWithSameRole);
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
